Code refactor with useConnectMCAccount.
